토큰이 만료되었을 경우 기기에 저장된 로그인 정보를 지우도록 변경

+++ app/src/main/java/com/youth/farm_volunteering/data/LikeData.kt
@@ -3,7 +3,7 @@ package com.youth.farm_volunteering.data
 class LikeData {
     var idx: Int? = null
     var price: Int? = null
-    var star: Int? = null
+    var star: Float? = null
     var period: String? = null
     var name: String? = null
     var isBooked: Int? = null
